requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
Pillow>=10.0.0
midiutil>=1.2.1
